requirements.txt
pytorch=1.0.1
torchvision==0.2.2
matplotlib==3.1.0
numpy==1.16.4
scikit-image
pandas
sklearn